+++ S5T2_SpringBoot_Dice_Roller_Microservice_Game/src/main/java/com/plopez/diceroller/microservice/game/model/entity/Game.java
@@ -18,8 +18,8 @@ public class Game {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private int playerId;
-    private byte die1;
-    private byte die2;
-    private byte result;
+    private int die1;
+    private int die2;
+    private int result;
 
 }
